Expose flutter.jar to module consumer. Simplify API
